+++ operate/client/src/modules/queries/incidents/useGetIncidentsByElementInstance.ts
@@ -6,17 +6,20 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import type {QueryIncidentsResponseBody} from '@camunda/camunda-api-zod-schemas/8.8';
+import type {QueryElementInstanceIncidentsResponseBody} from '@camunda/camunda-api-zod-schemas/8.8';
 import {useQuery} from '@tanstack/react-query';
+import {searchIncidentsByElementInstance} from 'modules/api/v2/incidents/searchIncidentsByElementInstance';
 import {queryKeys} from '../queryKeys';
 
 type QueryOptions<T> = {
   enabled?: boolean;
   enablePeriodicRefetch?: boolean;
-  select?: (result: QueryIncidentsResponseBody) => T;
+  select?: (result: QueryElementInstanceIncidentsResponseBody) => T;
 };
 
-const useGetIncidentsByElementInstance = <T = QueryIncidentsResponseBody>(
+const useGetIncidentsByElementInstance = <
+  T = QueryElementInstanceIncidentsResponseBody,
+>(
   elementInstanceKey: string,
   options?: QueryOptions<T>,
 ) => {
@@ -27,9 +30,13 @@ const useGetIncidentsByElementInstance = <T = QueryIncidentsResponseBody>(
     enabled: options?.enabled ?? !!elementInstanceKey,
     select: options?.select,
     queryFn: async () => {
-      throw new Error(
-        'Waiting for implementation: https://github.com/camunda/camunda/issues/38806',
-      );
+      const {response, error} =
+        await searchIncidentsByElementInstance(elementInstanceKey);
+      if (response !== null) {
+        return response;
+      }
+
+      throw error;
     },
   });
 };